requirements.txt
fastapi
uvicorn
noise
pytest
flake8
isort
httpx
pyyaml